+++ R/funs-date.R
@@ -87,6 +87,15 @@ pl_date_lubridate <- function(x, ...) {
   pl$date(year = year, month = month, day = day)
 }
 
+pl_today_lubridate <- function(tzone = "") {
+  pl$lit(lubridate::today(tzone = tzone))
+}
+
+pl_now_lubridate <- function(tzone = "") {
+  pl$lit(lubridate::now(tzone = tzone))
+}
+
+
 # Durations --------------------------------------
 
 # TODO: annoying to parse `...` because of partial matching of args, e.g